requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain_groq
pandas
numpy>=2.1.0
python-dotenv